requirements.txt
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fake-useragent==1.0.1
fastapi==0.87.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
loguru==0.6.0
multidict==6.0.2
pydantic==1.10.2
python-dotenv==0.21.0
python-socks==2.0.3
requests==2.28.1
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
yarl==1.8.1
